+++ tests/debug.rs
@@ -10,24 +10,22 @@ fn error_debug() {
         format!("{:?}", ex("hello\"", false)),
         r#"Ok("hello\"")"#.to_owned()
     );
+    // the long part tests that it puts location on another line
     let tmp = ex("hello", true)
         .stack()
         .stack_locationless()
-        .stack_err("test")
+        .stack_err("test long ___________________________________________________________________")
         .stack_err_locationless(ron::from_str::<bool>("invalid").unwrap_err())
         .stack_err_locationless(Box::new(ron::from_str::<bool>("invalid").unwrap_err()));
     println!("{tmp:?}");
     assert_eq!(
-        format!("{:?}", tmp),
-        r#"Err(Error { stack: [
-1:1: Expected boolean
-1:1: Expected boolean
-Location { file: "tests/debug.rs", line: 16, col: 10 },
-test
-Location { file: "tests/debug.rs", line: 14, col: 10 },
-Location { file: "tests/debug.rs", line: 39, col: 13 },
-hello
-] })"#
+        format!("{}", tmp.unwrap_err()),
+        r#"1:1: Expected boolean,
+1:1: Expected boolean,
+test long ___________________________________________________________________
+at tests/debug.rs 17:10,
+at tests/debug.rs 15:10,
+hello at tests/debug.rs 37:13"#
             .to_owned()
     );
 }