+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - `stack_err*` functions that take closures were renamed to `stack_err_with*`. The new `stack_err` now takes an expression directly.
 - More swap-in compatibility with `eyre` and `anyhow`
 - The error type now implements `core::error::Error`
+- The error type deliberately does not implement any `From` impls so that all initial `?`s need to be covered
+- Introduced `bail!` macro for easily creating and returning errors
+- Discovered an incredibly concise yet more distinctive format. `Debug` mode uses terminal styling.
 - The crate is now unconditionally `no_std`
 - Moved `stacked_get` and `stacked_get_mut` to this crate
 - many small improvements